Crate pezkuwi_sdk_docs

Pezkuwi SDK Docs

The Pezkuwi SDK Developer Documentation.

This crate is a minimal, always-accurate and low level source of truth about Pezkuwi-SDK. For more high level docs, please go to docs.pezkuwi.com.

Getting Started

We suggest the following reading sequence:

Start by learning about pezkuwi_sdk, its structure and context.

Then, head over to the guides. This modules contains in-depth guides about the most important user-journeys of the Pezkuwi SDK.

Whilst reading the guides, you might find back-links to reference_docs.

external_resources for a list of 3rd party guides and tutorials.

Finally, https://paritytech.github.io is the parent website of this crate that contains the list of further tools related to the Pezkuwi SDK.

Information Architecture

This section paints a picture over the high-level information architecture of this crate.

flowchart TD
    dot[pezkuwichain.io] --> devhub[pezkuwi_sdk_docs]

    devhub --> pezkuwi_sdk
    devhub --> reference_docs
    devhub --> guides
	devhub --> external_resources

    pezkuwi_sdk --> substrate
    pezkuwi_sdk --> frame
    pezkuwi_sdk --> xcm
	pezkuwi_sdk --> templates

Modules

external_resources

A list of external resources and learning material about Pezkuwi SDK.

guides

In-depth guides about the most common components of the Pezkuwi SDK. They are slightly more high level and broad than reference_docs.

meta_contributing

Meta information about this crate, how it is built, what principles dictates its evolution and how one can contribute to it.

pezkuwi_sdk

An introduction to the Pezkuwi SDK. Read this module to learn about the structure of the SDK, the tools that are provided as a part of it, and to gain a high level understanding of each.

reference_docs

Reference documents covering in-depth topics across the Pezkuwi SDK. It is suggested to read these on-demand, while you are going through the guides or other content.