pezkuwi_sdk_docs

Module guides

In-depth guides about the most common components of the Pezkuwi SDK. They are slightly more high level and broad than reference_docs.

Pezkuwi SDK Docs Guides

This crate contains a collection of guides that are foundational to the developers of Pezkuwi SDK. They are common user-journeys that are traversed in the Pezkuwi ecosystem.

The main user-journey covered by these guides is:

your_first_pallet, where you learn what a FRAME pallet is, and write your first application logic.

your_first_runtime, where you learn how to compile your pallets into a WASM runtime.

your_first_node, where you learn how to run the said runtime in a node.

By this step, you have already launched a full Pezkuwi-SDK-based blockchain!

Once done, feel free to step up into one of our templates: crate::pezkuwi_sdk::templates.

Other guides are related to other miscellaneous topics and are listed as modules below.

Modules

async_backing_guide

How to enable Async Backing on teyrchain projects that started in 2023 or before.

enable_elastic_scaling

How to enable elastic scaling on a teyrchain.

enable_metadata_hash

How to enable metadata hash verification in the runtime.

enable_pov_reclaim

How to enhance a given runtime and node to be cumulus-enabled, run it as a teyrchain and connect it to a relay-chain. How to enable storage weight reclaiming in a teyrchain node and runtime.

handling_teyrchain_forks

How to parameterize teyrchain forking in relation to relay chain forking.

your_first_node

Running the given runtime with a node. No specific consensus mechanism is used at this stage.

your_first_pallet

Write your first simple pallet, learning the most most basic features of FRAME along the way.

your_first_runtime

Write your first real runtime, compiling it to WASM.