pezkuwi_sdk_docs::reference_docs::frame_runtime_types::pallet_with_specific_runtime_call

Trait Config

pub trait Config: Config {
    type RuntimeCall: IsSubType<Call<Self>>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. A pallet that wants to further narrow down what RuntimeCall is.

Required Associated Types

type RuntimeCall: IsSubType<Call<Self>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.